Refactor code to split it into a new log.go file
